+++ db/migrations/002_publications_table_add_index_on_cbe_number_column.rb
@@ -1,0 +1,7 @@
+Sequel.migration do
+  change do
+    alter_table(:publications) do
+      add_index :cbe_number, name: "publications_cbe_number_index"
+    end
+  end
+end